update poetry.lock to update dependencies
This commit updates httplib2 to 0.18.0 by running "poetry lock"
command (see dependabot PR #289).

PR #289 updated the httplib2
dependency to 0.18.*, and the automated PR only had changes to the poetry.lock file, I’m not sure why that was the case - perhaps because it is a dependency of “google-auth-httplib2” package, which we *do* declare in our pyproject.toml.

Dependabot's PR had a merge conflict
with the poetry.lock file, running poetry lock resolved that.