Revert to original newrelic agent version; bump version
Originally thought that bumping the newrelic version would be a minor
upgrade that was small enough to be commited straight to master.

On second thought, I want to break that into it's own PR so I can test
that more fully on staging.

This commit reverts the change made in v1.2.23, and bumps the version to
1.2.24.